Rename the parser
